extract addon to packages
