+++ src/bootstrap.js
@@ -9,6 +9,6 @@ import 'react-app-polyfill/ie9';
 import './packages/react-google-translate-fix';
 
 /**
- * Initialise i18n
+ * Initialise i18next
  */
-import './common/services/i18n';
+import './common/services/i18next';
