+++ src/features/auth/ducks.js
@@ -4,7 +4,7 @@ import { push } from 'connected-react-router';
 import { createActionCreator } from '../../packages/redux-helpers';
 
 import { rootPath } from '../../config';
-import { t } from '../../common/services/i18n';
+import { t } from '../../common/services/i18next';
 import { loginActions } from '../../common/services/user';
 import { message } from '../../common/components';
